У меня каким то образом потерялась связь локального проекта с гитхабом. Поэтому в этом комите видны не все изменеия, и к тому пришлось создать новый репозиторий. Если получится, попробую всё таки склонировать проектс гитхаба, отразить в нём все изменеия и снова запушить на гитхаб.
