+++ app/src/main/java/com/softdesign/school/ui/fragments/ProfileFragment.java
@@ -1,0 +1,24 @@
+package com.softdesign.school.ui.fragments;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.accessibility.AccessibilityEvent;
+
+
+
+
+public class ProfileFragment extends Fragment {
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState){
+     View convertView = inflater.inflate(R.layout.fragment_profile, null, false);
+    getActivity().setTitle(getResources().getString(R.string.drawer_profile));
+
+       return convertView;
+    }
+}
