+++ app/src/main/res/values/colors.xml
@@ -4,13 +4,12 @@
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
 
-    <color name="colorRed">#FF0000</color>
-    <color name="colorDarkRed">#800000</color>
-    <color name="colorGreen">#00FF00</color>
-    <color name="colorDarkGreen">#008000</color>
-    <color name="colorBlue">#0000FF</color>
-    <color name="colorDarkBlue">#000080</color>
-
+    <color name="red">#FF0000</color>
+    <color name="dark_red">#800000</color>
+    <color name="green">#009900</color>
+    <color name="dark_green">#007000</color>
+    <color name="blue">#0000FF</color>
+    <color name="dark_blue">#000080</color>
 
     <color name="icon_color">#ffffff</color>
 </resources>
