+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,17 @@
 <resources>
     <string name="app_name">SDApplication</string>
+
+    <string name="drawer.profile">Профиль</string>
+    <string name="drawer.contacts">Контакты</string>
+    <string name="drawer.tasks">Задачи</string>
+    <string name="drawer.team">Команда</string>
+    <string name="drawer.setting">Настройки</string>
+
+    <string name="fragment.contacts.description">Здесь будут контакты</string>
+    <string name="fragment.profile.description">Здесь будет профиль</string>
+    <string name="fragment.team.description">Здесь будет команда</string>
+    <string name="fragment.tasks.description">Здесь будут задачи</string>
+    <string name="fragment.settings.description">Здесь будут настройки</string>
+
+
 </resources>
